refactor: Move scoring functions into separate scoring module
